+++ antargis/trunk/src/entity.h
@@ -143,7 +143,7 @@ class AntEntity:public AGRubyObject
     virtual void newRestJob(float pTime);
     virtual void newFetchJob(int p,AGVector2 &pTarget,const std::string &pWhat);
     virtual void newFetchJob(int p,AntEntity *pTarget,const std::string &pWhat);
-    virtual void newMoveJob(int p,const AGVector2 &pTarget,int pnear=0);
+    virtual void newMoveJob(int p,const AGVector2 &pTarget,float pnear=0.0);
     virtual void newFightJob(int p,AntEntity *target);
     virtual void delJob();
